--- a/lib/cocktail/schedule_state.ex
+++ b/lib/cocktail/schedule_state.ex
@@ -3,6 +3,8 @@ defmodule Cocktail.ScheduleState do
 
   alias Cocktail.{Schedule, RuleState, Span}
 
+  require Logger
+
   @type t :: %__MODULE__{
               recurrence_rules: [RuleState.t],
               recurrence_times: [Cocktail.time],
@@ -35,6 +37,8 @@ defmodule Cocktail.ScheduleState do
 
   @spec next_time(t) :: {Cocktail.occurrence, t}
   def next_time(%__MODULE__{} = state) do
+    Logger.debug(fn -> "~> starting time: #{state.current_time}" end)
+
     {time, remaining_rules} = next_time_from_recurrence_rules(state)
     {time, remaining_times} = next_time_from_recurrence_times(state.recurrence_times, time)
     {is_exception, remaining_excpetions} = apply_exception_time(state.exception_times, time)
@@ -44,6 +48,8 @@ defmodule Cocktail.ScheduleState do
         if is_exception do
           next_time(state)
         else
+          Logger.debug(fn -> "=> occurrence: #{occurrence}" end)
+
           {occurrence, state}
         end
       nil ->