… lib/cocktail/validation/interval.ex
… lib/cocktail/validation/minute_of_hour.ex
… lib/cocktail/validation/schedule_lock.ex
… lib/cocktail/validation/second_of_minute.ex
… lib/cocktail/validation/shift.ex
… lib/cocktail/validation/time_of_day.ex
… test/cocktail/builder/string_test.exs
… test/cocktail/parser/i_calendar_test.exs
… test/cocktail/parser/json_test.exs
… test/cocktail/reversibility_test.exs
… test/cocktail/rule_test.exs
… test/cocktail/schedule_test.exs
… test/cocktail/span_test.exs
… test/cocktail/validation/time_of_day_test.exs