TODO:
*   [x] Before and after benchee comparison of various schedules
*   [x] Look into not having to re-run the entire list of validations for a rule just because one validation changed the time.  Some should be trusted to not change it again.
*   [x] Look into using balanced binary tree for list-based validations to reduce the number of checks needed
*   [x] Support parsing and building iCalendar with time of day options