@doughsay 2017-09-22
# Master

```
##### With input bad schedule #####
Name                 ips        average  deviation         median
first 10          151.43        6.60 ms    ±12.55%        6.25 ms
first 100          13.86       72.17 ms     ±5.00%       70.80 ms
first 1000          1.38      723.14 ms     ±1.29%      719.99 ms

Comparison:
first 10          151.43
first 100          13.86 - 10.93x slower
first 1000          1.38 - 109.50x slower

##### With input complex schedule #####
Name                 ips        average  deviation         median
first 10          848.85        1.18 ms     ±9.63%        1.15 ms
first 100          90.44       11.06 ms     ±6.84%       10.85 ms
first 1000          9.14      109.40 ms     ±3.60%      107.73 ms

Comparison:
first 10          848.85
first 100          90.44 - 9.39x slower
first 1000          9.14 - 92.86x slower

##### With input simple schedule #####
Name                 ips        average  deviation         median
first 10         7292.55       0.137 ms    ±12.76%       0.134 ms
first 100         683.72        1.46 ms    ±12.35%        1.41 ms
first 1000         70.71       14.14 ms     ±8.85%       13.66 ms

Comparison:
first 10         7292.55
first 100         683.72 - 10.67x slower
first 1000         70.71 - 103.13x slower
```

# Performance branch

```
##### With input bad schedule #####
Name                 ips        average  deviation         median
first 10          556.34        1.80 ms    ±10.14%        1.74 ms
first 100          47.94       20.86 ms     ±5.87%       20.42 ms
first 1000          4.73      211.53 ms     ±5.94%      207.99 ms

Comparison:
first 10          556.34
first 100          47.94 - 11.61x slower
first 1000          4.73 - 117.68x slower

##### With input complex schedule #####
Name                 ips        average  deviation         median
first 10         12.53 K       79.84 μs    ±14.88%       77.00 μs
first 100         1.32 K      757.14 μs     ±9.60%      736.00 μs
first 1000       0.129 K     7748.55 μs    ±10.35%     7401.50 μs

Comparison:
first 10         12.53 K
first 100         1.32 K - 9.48x slower
first 1000       0.129 K - 97.06x slower

##### With input simple schedule #####
Name                 ips        average  deviation         median
first 10         12.13 K       82.41 μs    ±13.22%       79.00 μs
first 100         1.12 K      893.35 μs    ±10.83%      861.00 μs
first 1000       0.116 K     8630.30 μs    ±10.67%     8231.50 μs

Comparison:
first 10         12.13 K
first 100         1.12 K - 10.84x slower
first 1000       0.116 K - 104.72x slower
```