@doughsay 2017-09-22
Discovered truths:
*   All validations for a rule must give a "pass" to the time before it is emitted.
*   When a validation fails, it moves the time to the next time that "passes"
*   If a validation moves the time, all other validations at the same level or above have their "pass" status cleared and must be re-checked.

By following these rules, I can reduce the number of validation checks performed in the loop by a decent amount; should increase performance even more.