@doughsay 2017-09-26
# Final comparison

```
##### With input 01 complex schedule #####
Name                 ips        average  deviation         median
first 10         13.57 K       73.67 μs    ±22.25%       70.00 μs
first 100         1.45 K      687.56 μs    ±14.00%      650.00 μs
first 1000       0.146 K     6855.83 μs    ±12.99%     6484.50 μs

Comparison:
first 10         13.57 K
first 100         1.45 K - 9.33x slower
first 1000       0.146 K - 93.06x slower

##### With input 02 simple schedule #####
Name                 ips        average  deviation         median
first 10         12.13 K       82.42 μs    ±15.00%       79.00 μs
first 100         1.14 K      875.77 μs    ±11.12%      847.00 μs
first 1000       0.118 K     8447.83 μs    ±10.41%     8100.50 μs

Comparison:
first 10         12.13 K
first 100         1.14 K - 10.63x slower
first 1000       0.118 K - 102.50x slower

##### With input 03 bad schedule #####
Name                 ips        average  deviation         median
first 10          609.85        1.64 ms    ±10.06%        1.59 ms
first 100          51.56       19.39 ms     ±6.65%       18.85 ms
first 1000          5.09      196.64 ms     ±4.53%      193.57 ms

Comparison:
first 10          609.85
first 100          51.56 - 11.83x slower
first 1000          5.09 - 119.92x slower

##### With input 04 fixed bad schedule #####
Name                 ips        average  deviation         median
first 10         10.13 K       98.73 μs    ±13.13%       96.00 μs
first 100         1.05 K      955.27 μs    ±13.50%      910.00 μs
first 1000       0.105 K     9533.37 μs     ±8.47%     9223.00 μs

Comparison:
first 10         10.13 K
first 100         1.05 K - 9.68x slower
first 1000       0.105 K - 96.56x slower
```

Comparing our "bad schedule" on master to our "fixed bad schedule" on the latest update; getting the first 1000 occurrences:

```
first 1000          1.38      723.14 ms     ±1.29%      719.99 ms
first 1000       0.105 K     9533.37 μs     ±8.47%     9223.00 μs
```

From ~720ms to ~9.2ms; that's a **78.2x improvement.**